+++ Cargo.toml
@@ -29,7 +29,6 @@ members = [
   "commons/zenoh-task",
   "commons/zenoh-util",
   "commons/zenoh-runtime",
-  "commons/zenoh-runtime/zenoh-runtime-derive",
   "examples",
   "io/zenoh-link",
   "io/zenoh-link-commons",
@@ -203,7 +202,6 @@ zenoh-link-commons = { version = "0.11.0-dev", path = "io/zenoh-link-commons" }
 zenoh = { version = "0.11.0-dev", path = "zenoh", default-features = false }
 zenoh-runtime = { version = "0.11.0-dev", path = "commons/zenoh-runtime" }
 zenoh-task = { version = "0.11.0-dev", path = "commons/zenoh-task" }
-zenoh-runtime-derive = { version = "0.11.0-dev", path = "commons/zenoh-runtime/zenoh-runtime-derive" }
 
 [profile.dev]
 debug = true
