+++ commons/zenoh-runtime/Cargo.toml
@@ -16,9 +16,9 @@ description = { workspace = true }
 libc = { workspace = true }
 ron = { workspace = true }
 serde = { workspace = true }
-zenoh-runtime-derive = { workspace = true }
 futures = { workspace = true }
 lazy_static = { workspace = true }
 zenoh-result = { workspace = true, features = ["std"] }
 zenoh-collections = { workspace = true, features = ["std"] }
+zenoh-macros = { workspace = true }
 tokio = { workspace = true, features = ["fs", "io-util", "macros", "net", "rt-multi-thread", "sync", "time"] }
