--- commons/zenoh-runtime/zenoh-runtime-derive/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "zenoh-runtime-derive"
-version = { workspace = true }
-repository = { workspace = true }
-homepage = { workspace = true }
-authors = {workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-categories = { workspace = true }
-description = "Internal crate for zenoh."
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-syn = { workspace = true }
-quote = { workspace = true }
-proc-macro2 = { workspace = true }
-
-[lib]
-proc-macro = true
